Add files via upload
Created a directory called courses and in page.tsx it will list the information of course requirements for majors such as Computer Science, Multimedia Computing, and Information Systems, minors such as Computer Science, Data Science, and Multimedia Computing, and some resources in preparing for the course. The index.tsx is providing a section to view the extensive details of the course. Then I created a sub directory called [courseCode] which has a file called page.tsx which will the details of the course such as the prerequisite, number of credits, and number of hours in the course.